chore: add simple livez endpoint
This will be used for liveness probe on the sidecar and will make
sure that the sidecar is accepting calls, rather than verifying that
the e2e check is passing.

This should address an issue where both containers crashloop at the
same time, which caused the check to fail constantly.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>