refactor: check runs in loop
To avoid having a GET call on an endpoint trigger the check
(side effect), this approach runs the test in loop and uses a shared
volume to store the results.

For liveness, Smee is verifying the age and result of the check,
while the sidecar is only verifying the age of the test. In other
words, Smee verifies that the setup is alive, while the sidecar only
verifies that the check is running.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>