fix: sidecar fails to rewrite scripts
The sidecar is writing the scripts to the shared volume on startup
and it's making them read-only.

We didn't take into account the scenario in which the sidecar
crashes and the shared volume persisted. In this case, it will fail
to overwrite the read-only scripts.

This change should address the issue.

Signed-off-by: Yftach Herzog <yherzog@redhat.com>